Remove the cbr0 definition from the formula. This may seem strange, so here are the two reason why:
1. This interface definition causes Ubuntu systems to hang for 2-4 minutes on boot, during cloud-init’s startup.
2. It has no clear purpose, at least on Ubuntu systems.

The cbr0 bridge, much like the docker0 bridge, is created when necessary. The same way it doesn’t make any sense for docker0 to be defined in /etc/network/interfaces, I don’t believe it makes any sense to define cbr0 in /etc/network/interfaces. Removing this definition had the following effects:

1. It removes the system boot issue.
2. The cbr0 interface is now only present after a pod is assigned to the node. If the node is running with no pods, no cbr0 interface is created.

Both of these behaviors make sense. The only one that could be a little confusing is 2.