Adds the certs.generate and certs.present states and removes the certs.configured and certs.view states.
certs.generate generates the certificates and prints it in a format that can be placed into the pillar with a cut and paste.
certs.present places the certificates stored in the pillar in the appropriate places, along with adding it to the system certificates list.

certs.removed was modified to remove the certificate from the system certificates list by running the update-ca-certificates command

The cert folder was renamed to certs since it manages multiple certificates

The README.md was updated to reflect these changes.