adds the ip_helper module for the kubernetes formula so that add-on services that need to be in the cluster service CIDR can have pillar data derived from that CIDR. The services that require an ip address that are now derived using the ip_helper module are the following: dns_server - the kube-dns server ip address registry_server - the docker registry server ip address
These can still be manually overridden, but you'll want to be careful to make sure the ip addresses remain in the service_cluster_ip_range.

Updates map.jinja file to get the derived values.
Updates default.yml to remove registry_server and dns_server values, adds a note to go check the map.jinja if you want to see how this works.
Updates the README for the registry server.